requirements.txt
faster-whisper==1.0.3
groq==0.11.0
python-docx==1.1.2
